requirements.txt
google-genai>=0.3.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
opencv-python>=4.8.0
moviepy>=1.0.3
numpy>=1.24.0
typing-extensions>=4.7.0
